Attempt to fix the flaky Wai.Middleware.LoggingSpec test case
Log messages from the previous property run were sometimes mixed with
the results for the current run.

This happened much more on Windows, but I found I could also reproduce
it with about 25-50 test runs on Linux.

I initially thought it was due to some buffering in
iohk-monitoring. But that was wrong - traceInTVarIO is actually a
really simple trace transformer (pretty cool actually!).

Perhaps what is happening is that the captured log list isn't fully
evaluated before the test case finishes.